Adds handling for case when the regex pattern can't identify a match
This commit updates a closure responsible for translating the match for
a pattern in the status page to a metric so that a default value can be
provided if no match is found.

If there are no failed datasets, the closure would raise an Exception.
Now, the closure is provided a default value of 0, so that if there are
no failed datasets, the metric produced is reported as 0.